feat:【声音样本库】添加用户声音库表，备份数据库

+++ SoundMentor-pojo/src/main/java/com/soundmentor/soundmentorpojo/DTO/basic/IdParam.java
@@ -22,5 +22,5 @@ public class IdParam {
      * @see String
      */
     @NotNull(message = "ID不能为空")
-    protected Long id;
+    protected Integer id;
 }
